+++ clown_sort/files/pdf_file.py
@@ -11,11 +11,12 @@ from pypdf.errors import DependencyError, EmptyFileError
 from rich.console import Console
 from rich.panel import Panel
 
-from clown_sort.config import check_for_pymupdf, log_optional_module_warning
+from clown_sort.config import (MIN_PDF_SIZE_TO_LOG_PROGRESS_TO_STDERR, Config, check_for_pymupdf,
+     log_optional_module_warning)
 from clown_sort.files.image_file import ImageFile
 from clown_sort.files.sortable_file import SortableFile
 from clown_sort.util.logging import log
-from clown_sort.util.rich_helper import WARNING, console
+from clown_sort.util.rich_helper import WARNING, console, stderr_console
 
 MAX_DISPLAY_HEIGHT = 600
 SCALE_FACTOR = 0.4
@@ -30,33 +31,46 @@ class PdfFile(SortableFile):
             return self._extracted_text
 
         log.debug(f"Extracting text from '{self.file_path}'...")
-        console_buffer = Console(file=io.StringIO())
         extracted_pages = []
 
         try:
             pdf_reader = PdfReader(self.file_path)
+            page_count = len(pdf_reader.pages)
+            log.debug(f"PDF Page count: {page_count}")
 
-            for page_number, page in enumerate(pdf_reader.pages):
-                console_buffer.print(Panel(f"PAGE {page_number + 1}", padding=(0, 15), expand=False))
-                console_buffer.print(page.extract_text().strip())
+            for page_number, page in enumerate(pdf_reader.pages, start=1):
+                self._log_to_stderr(f"Parsing page {page_number}...")
+                page_buffer = Console(file=io.StringIO())
+                page_buffer.print(Panel(f"PAGE {page_number}", padding=(0, 15), expand=False))
+                page_buffer.print(page.extract_text().strip())
+                image_number = 1
 
-                for image_number, image in enumerate(page.images, start=1):
-                    image_name = f"Page {page_number + 1}, Image {image_number}"
-                    console_buffer.print(Panel(image_name, expand=False))
-                    image_obj = Image.open(io.BytesIO(image.data))
-                    image_text = ImageFile.extract_text(image_obj, f"{self.file_path} ({image_name})")
-                    console_buffer.print((image_text or '').strip())
+                # Extracting images is a bit fraught (lots of PIL and pypdf exceptions have come from here)
+                try:
+                    for image_number, image in enumerate(page.images, start=1):
+                        image_name = f"Page {page_number}, Image {image_number}"
+                        self._log_to_stderr(f"   Processing {image_name}...")
+                        page_buffer.print(Panel(image_name, expand=False))
+                        image_obj = Image.open(io.BytesIO(image.data))
+                        image_text = ImageFile.extract_text(image_obj, f"{self.file_path} ({image_name})")
+                        page_buffer.print((image_text or '').strip())
+                except (NotImplementedError, OSError, TypeError, ValueError) as e:
+                    stderr_console.print(f"WARNING: {type(e).__name__}: {e} while parsing embedded image {image_number} on page {page_number}...")
+                    stderr_console.print_exception()
 
-                page_text = console_buffer.file.getvalue()
-                log.debug(page_text)
+                page_text = page_buffer.file.getvalue()
                 extracted_pages.append(page_text)
+                log.debug(page_text)
+
+                if Config.print_when_parsed:
+                    print(f"{page_text}")
         except DependencyError:
             log_optional_module_warning('pdf')
         except EmptyFileError:
             log.warn("Skipping empty file!")
 
-        self.text_extraction_attempted = True
         self._extracted_text = "\n\n".join(extracted_pages).strip()
+        self.text_extraction_attempted = True
         return self._extracted_text
 
     def thumbnail_bytes(self) -> Optional[bytes]:
@@ -94,6 +108,13 @@ class PdfFile(SortableFile):
             console.print(msg)
 
         return type(self).is_presentable_in_popup
+
+    def _log_to_stderr(self, msg: str) -> None:
+        """When parsing very large PDFs it can be useful to log progress and other messages to STDERR."""
+        if self.file_size() < MIN_PDF_SIZE_TO_LOG_PROGRESS_TO_STDERR:
+            return
+
+        stderr_console.print(msg)
 
     def __repr__(self) -> str:
         return f"PdfFile('{self.file_path}')"